requirements.txt
boto3==1.18.21
botocore==1.21.21
jmespath==0.10.0
mysql-connector-python==8.0.26
numpy==1.19.5
pandas==1.1.5
protobuf==3.17.3
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
schedule==1.1.0
six==1.16.0
urllib3==1.26.6
